// src/clr-angular/utils/drag-and-drop/interfaces/draggable-event.ts
/**
 * Copyright (c) 2016-2017 VMware, Inc. All Rights Reserved.
 * This software is released under MIT license.
 * The full license information can be found in LICENSE in the root directory of this project.
 */
import {DragPosition} from "./client-position";
import {Draggable} from "./draggable";

export interface DraggableEvent {
    draggable: Draggable;
    dragPosition: DragPosition;
    dragAndDropGroup: string;
    ghostAnchorPosition?: DragPosition;
    data: any;
}